wip: Fix build errors with 'libcu8' by comparing PBAT_USE_LIBCU8 to '1' to see if it is enabled

+++ libpbat/thread/pBat_Thread.c
@@ -143,7 +143,7 @@ LIBPBAT WINAPI DWORD _pBat_BeginThreadTrampoline(LPVOID data)
 
     free(s);
 
-    return (DWORD)ret;
+    return (DWORD)(size_t)ret;
 }
 
 LIBPBAT int  pBat_BeginThread(THREAD* lpThId, void*(*pFunc)(void*), int iMemAmount, void* arg)
